+++ src/components/Layout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
 
-import Header from '@/components/Header/Header';
+import Header from '@/components/Header';
 
 function Layout() {
   return (
